+++ src/Core/Dfe.Complete.Application/Projects/Models/ListProjectsByMonthResultModel.cs
@@ -1,7 +1,5 @@
-using Dfe.Complete.Domain.Entities;
 using Dfe.Complete.Domain.Enums;
 using Dfe.Complete.Domain.ValueObjects;
-using Dfe.Complete.Utils;
 
 namespace Dfe.Complete.Application.Projects.Models;