+++ src/Core/Dfe.Complete.Application/Projects/Queries/ListAllProjects/ListAllProjectsByRegions.cs
@@ -4,6 +4,7 @@ using Dfe.Complete.Application.Projects.Models;
 using Dfe.Complete.Domain.Enums;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 namespace Dfe.Complete.Application.Projects.Queries.ListAllProjects;
 
@@ -12,7 +13,7 @@ public record ListAllProjectsByRegionQuery(
     ProjectType? Type)
     : IRequest<Result<List<ListAllProjectsByRegionsResultModel>>>;
 
-public class ListAllProjectsByRegionQueryHandler(IListAllProjectsQueryService listAllProjectsQueryService)
+public class ListAllProjectsByRegionQueryHandler(IListAllProjectsQueryService listAllProjectsQueryService, ILogger<ListAllProjectsByRegionQueryHandler> logger)
     : IRequestHandler<ListAllProjectsByRegionQuery, Result<List<ListAllProjectsByRegionsResultModel>>>
 
 {
@@ -33,12 +34,14 @@ public class ListAllProjectsByRegionQueryHandler(IListAllProjectsQueryService li
                         ConversionsCount: group.Count(item => item.Project?.Type == ProjectType.Conversion),
                         TransfersCount: group.Count(item => item.Project?.Type == ProjectType.Transfer)
                     ))
+                .OrderBy(item => item.Region.ToString())
                 .ToList();
 
             return Result<List<ListAllProjectsByRegionsResultModel>>.Success(projectsResultModel);
         }
         catch (Exception e)
         {
+            logger.LogError(e, "Exception for {Name} Request - {@Request}", nameof(ListAllProjectsByRegionQueryHandler), request);
             return Result<List<ListAllProjectsByRegionsResultModel>>.Failure(e.Message);
         }
     }