+++ src/Frontend/Dfe.Complete/Pages/Search/SearchProjects.cshtml.cs
@@ -1,14 +1,14 @@
 using Dfe.Complete.Application.Projects.Models;
-using Dfe.Complete.Application.Projects.Queries.CountAllProjects;
 using Dfe.Complete.Application.Projects.Queries.SearchProjects;
-using Dfe.Complete.Domain.Enums;
 using Dfe.Complete.Pages.Pagination;
 using Dfe.Complete.Pages.Projects.List;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Dfe.Complete.Pages.Search
 {
+    [Authorize]
     public class SearchProjectsModel(ISender sender) : AllProjectsModel("")
     {
         [BindProperty(SupportsGet = true, Name = "query")]
@@ -36,17 +36,14 @@ namespace Dfe.Complete.Pages.Search
             }
             else
             {
-                var countProjectQuery = new CountAllProjectsQuery(ProjectState.Active, null, AssignedToState.AssignedOnly, Query);
-                var countResponse = await sender.Send(countProjectQuery);
-
-                var searchProjectsQuery = new SearchProjectsQuery(ProjectState.Active, Query)
+                var searchProjectsQuery = new SearchProjectsQuery(Query)
                 {
                     Page = PageNumber - 1,
                     Count = PageSize
                 };
                 var searchProjectsResponse = await sender.Send(searchProjectsQuery);
                 Projects = searchProjectsResponse.Value ?? [];
-                TotalResults = countResponse.Value;
+                TotalResults = searchProjectsResponse.ItemCount;
 
                 Pagination = new PaginationModel($"/search?query={Query}", PageNumber, TotalResults, PageSize);
             }