+++ CHANGELOG.md
@@ -22,13 +22,37 @@ Security in case of vulnerabilities.
 
 ## Unreleased  
 ### Added  
-- Manage local authorities under service support section.
+- `PolicyCheckTagHelper` added to conditionally hide elements based on policy
 
 ### Changed  
+- Navigation items previously hidden with `UserTabAccessHelper` now hide on policy
 
 ### Fixed  
 
-See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-05-16.272...main) for everything awaiting release
+### Removed
+- `UserTabAccessHelper` class is no longer required. Use policies instead
+
+### Security
+- Only correct user groups can now create projects
+
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-05-22.290...main) for everything awaiting release
+
+---
+
+## [1.5.3](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2025-05-22.290) - 2025-05-22
+### Added  
+ - Enabled error tracking via Application Insights.
+ - New route `/projects/team/unassigned`
+ - Your team projects "Unassigned" list (`/projects/team/unassigned`)
+
+### Changed  
+- Sort all projects by region list alphabetically
+
+### Fixed  
+- Note FK Ids are now required
+- Separated created and assigned users in project creation
+
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-05-16.272...production-2025-05-22.290) for everything in the release
 
 ---
 
@@ -41,6 +65,7 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 - Fixed identifying "Form A MAT" projects logic
 - Removed unnecessary `Assign To` filter while pulling projects from database.
 - Resolve accessibility issue causing app header to appear blue instead of white
+- Removed `Project Status` filter while pullling search results.  
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-05-14.254...production-2025-05-16.272) for everything in the release